DB refactor - running state

+++ composeApp/src/main/kotlin/com/example/a62550_foodapp/ui/shoppingList/ShoppingListPage.kt
@@ -138,7 +138,7 @@ fun ShoppingListPage(
             BackHandler { editListNameOverlay = false }
             EditShoppingListFormOverlay(
                 onDismiss = {editListNameOverlay = false},
-                onEdit = { id: Int, name: String ->
+                onEdit = { id: Long, name: String ->
                     shoppingListViewModel.editShoppingList(
                         id = id,
                         name = name
@@ -151,7 +151,7 @@ fun ShoppingListPage(
             BackHandler { deleteListOverlay = false }
             DeleteShoppingListFormOverlay(
                 onDismiss = {deleteListOverlay = false},
-                onDelete = { id: Int, name: String ->
+                onDelete = { id: Long, name: String ->
                     shoppingListViewModel.deleteShoppingList(
                         id = id,
                         name = name
@@ -304,7 +304,7 @@ private fun NewShoppingListFormOverlay(
 @Composable
 private fun EditShoppingListFormOverlay(
     onDismiss: () -> Unit,
-    onEdit: (Int, String) -> Unit,
+    onEdit: (Long, String) -> Unit,
     selectedShoppingList: ShoppingList?
 ) {
 
@@ -379,7 +379,7 @@ private fun EditShoppingListFormOverlay(
 @Composable
 private fun DeleteShoppingListFormOverlay(
     onDismiss: () -> Unit,
-    onDelete: (Int, String) -> Unit,
+    onDelete: (Long, String) -> Unit,
     selectedShoppingList: ShoppingList?,
     themeViewModel : ThemeViewModel = koinViewModel()
 ) {
